Task6 moving and scaling editing camera

+++ Assets/Scripts/FlappyBirdUpdated/LevelEditor.cs
@@ -1,7 +1,9 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Tilemaps;
+using UnityEngine.EventSystems;
 
 public class LevelEditor : MonoBehaviour
 {
@@ -45,11 +47,24 @@ public class LevelEditor : MonoBehaviour
     private void Update()
     {
         Vector3Int pos = currentTilemap.WorldToCell(cam.ScreenToWorldPoint(Input.mousePosition));
+
+        // Moving camera options
+        if (Input.GetKeyDown(KeyCode.D)) MovingCamera("right");
+        if (Input.GetKeyDown(KeyCode.A)) MovingCamera("left");
+        if (Input.GetKeyDown(KeyCode.W)) MovingCamera("up");
+        if (Input.GetKeyDown(KeyCode.S)) MovingCamera("down");
+
+        // Scaling camera options
+        if (Input.GetKeyDown(KeyCode.F)) ScalingCamera(-2);
+        if (Input.GetKeyDown(KeyCode.G)) ScalingCamera(2);
         
-        // Placing tile /w LMB
-        if (Input.GetMouseButton(0)) PlaceTile(pos);
-        // Deleting tile /w RMB
-        if (Input.GetMouseButton(1)) DeleteTile(pos);
+        if (!EventSystem.current.IsPointerOverGameObject())
+        {
+            // Placing tile /w LMB
+            if (Input.GetMouseButton(0)) PlaceTile(pos);
+            // Deleting tile /w RMB
+            if (Input.GetMouseButton(1)) DeleteTile(pos);
+        }
         
         // Selecting next tile in the list by pressing plus btn
         if (Input.GetKeyDown(KeyCode.Plus) || Input.GetKeyDown(KeyCode.KeypadPlus)) {
@@ -64,6 +79,43 @@ public class LevelEditor : MonoBehaviour
             if (_selectedTileIndex < 0) _selectedTileIndex = LevelManager.instance.tiles.Count - 1;
             Debug.Log($"Selected tile: {LevelManager.instance.tiles[_selectedTileIndex].name}");
         }
+    }
+
+    public void ScalingCamera(int scalingNumber)
+    {
+        var cameraSize = cam.GetComponent<Camera>().orthographicSize;
+
+        // if not (( camera size is minimal and trying to scale it down) or
+        //  ( if camera size is maximum and trying to scale it up))
+        if (!((cameraSize == 4 && scalingNumber < 0) ||
+               (cameraSize == 40 && scalingNumber > 0)))
+            cam.GetComponent<Camera>().orthographicSize += scalingNumber;
+    }
+
+    public void MovingCamera(string direction)
+    {
+        var cameraPosition = cam.transform.position;
+
+        switch (direction)
+        {
+            case "right":
+                cameraPosition.x += 5;
+                break;
+            case "left":
+                cameraPosition.x -= 5;
+                break;
+            case "up":
+                cameraPosition.y += 5;
+                break;
+            case "down":
+                cameraPosition.y -= 5;
+                break;
+            default:
+                Debug.Log("fuck");
+                break;
+        }
+
+        cam.transform.position = cameraPosition;
     }
 
     void PlaceTile(Vector3Int pos)
